YH-378: refactor log and reg forms, styling
