+++ src/entities/auth/model/types/auth.ts
@@ -1,5 +1,6 @@
 export interface Auth {
-	username: string;
+	// username: string;
+	email: string;
 	password: string;
 }
