--- src/features/registration/ui/RegisterForm/RegisterForm.module.css
@@ -1,83 +1,0 @@
-.wrapper {
-  max-width: 408px;
-}
-
-.form-wrapper {
-  margin-bottom: 40px;
-}
-
-.input-wrapper {
-  position: relative;
-  margin-bottom: 36px;
-}
-
-.label {
-  display: block;
-  margin-bottom: 8px;
-}
-
-.input {
-  position: relative;
-  padding-left: 20px;
-  border-radius: 68px;
-  background-color: var(--background-app);
-}
-
-.input input {
-  color: var(--text-color-primary)
-}
-
-.error {
-  position: absolute;
-  color: var(--color-red-600);
-  text-align: center;
-  line-height: normal;
-  font-size: var(--font-size-p-xs);
-  user-select: none;
-}
-
-.submit-button {
-  margin-bottom: 14px;
-  width: 100%;
-}
-
-.consent-wrapper {
-  display: flex;
-  align-items: center;
-}
-
-.consent-text {
-  margin-left: 12px;
-  line-height: normal;
-  font-size: var(--font-size-p-xs);
-}
-
-.consent-wrapper input[type="checkbox"] {
-  flex-shrink: 0;
-  width: 20px;
-  height: 20px;
-  appearance: none;
-  outline: none;
-  border: 2px solid var(--text-color-link-active);
-  border-radius: 4px;
-  transition: 0.3s ease;
-  cursor: pointer;
-}
-
-.checkbox:checked::before {
-  position: absolute;
-  top: -2px;
-  right: 1px;
-  color: var(--color-white-900);
-  font-size: var(--font-size-p-l);
-  content: "\2714";
-}
-
-.checkbox:checked {
-  position: relative;
-  background-color: var(--text-color-link-active);
-}
-
-.consent-wrapper input[type="checkbox"]:not(:checked):hover {
-  border-color: var(--color-purple-500)
-}
