+++ src/shared/ui/FormControl/ui/FormControl.tsx
@@ -1,18 +1,24 @@
 import { ReactNode, useMemo } from 'react';
-import { useController, Control, ControllerRenderProps, FieldValues } from 'react-hook-form';
+import { useController, Control, ControllerRenderProps, FieldValues, Path } from 'react-hook-form';
 import { FormControl as CustomControl } from 'yeahub-ui-kit';
 
 type ChildrenProps = Omit<ControllerRenderProps<FieldValues, string>, 'ref'>;
 
-interface contrProps {
-	name: string;
-	control: Control;
+interface FormControlProps<T extends FieldValues> {
+	name: Path<T>;
+	control: Control<T>;
 	label?: string;
 	className?: string;
 	children: (field: ChildrenProps, error: boolean) => ReactNode;
 }
 
-export const FormControl = ({ children, name, control, label, className }: contrProps) => {
+export const FormControl = <T extends FieldValues>({
+	children,
+	name,
+	control,
+	label,
+	className,
+}: FormControlProps<T>) => {
 	const {
 		field: { ref, ...fieldProps },
 		fieldState: { error },
