+++ src/shared/ui/InputPhone/ui/InputPhone.module.css
@@ -3,7 +3,7 @@
   align-items: center;
   padding: 8px;
   width: 100%;
-  min-height: 40px;
+  height: 48px;
   border: 1px solid var(--color-black-100);
   border-radius: 68px;
   font-family: var(--font-family);
@@ -25,24 +25,23 @@
   color: var(--text-color-primary);
 }
 
-.phone:focus-within {
-  caret-color: var(--color-purple-500);
-  box-shadow: var(--focus-shadow);
-  border-color: var(--color-purple-500);
-}
-
 .phone::placeholder {
   color: var(--text-color-more-lightest);
+  font-weight: var(--font-weight-400);
   font-size: var(--font-size-p-s);
 }
 
 .phone.error {
-  border-color: var(--color-red-800);
+  border-color: var(--color-red-700);
+  background-color: var(--background-block);
+}
+
+.phone:hover {
+  border-color: var(--palette-ui-purple-300);
 }
 
 .phone:focus-within.error {
   caret-color: var(--color-red-400);
-  box-shadow: var(--focus-shadow-error);
+  box-shadow: var(--focus-shadow);
+  border-color: var(--color-purple-500);
 }
-
-
